1 手势处理完成 2 大小屏幕切换完成 3 控制View处理完成

+++ Lib_Media/src/main/java/com/xm/lib/media/help/VolumeHelper.kt
@@ -2,6 +2,7 @@ package com.xm.lib.media.help
 
 import android.content.Context
 import android.media.AudioManager
+import android.util.Log
 
 
 class VolumeHelper {
@@ -16,7 +17,7 @@ class VolumeHelper {
          */
         fun getVolume(context: Context?): Float {
             val audioManager: AudioManager = context?.getSystemService(Context.AUDIO_SERVICE) as AudioManager
-            return (audioManager.getStreamVolume(AudioManager.STREAM_MUSIC) / audioManager.getStreamMaxVolume(AudioManager.STREAM_MUSIC)).toFloat()
+            return audioManager.getStreamVolume(AudioManager.STREAM_MUSIC).toFloat() / audioManager.getStreamMaxVolume(AudioManager.STREAM_MUSIC).toFloat()
         }
 
         /**
